feat(skill): 并发不足无豁免(删 no-work-after-audit escape)+ wakeup_plan→cli/filter-open(#162)
