+++ skills/codex-refactor-loop/scripts/concurrency_monitor.py
@@ -628,8 +628,9 @@ def tick() -> None:
     target = max(expected, floor)
     if actual < target:
         if dispatch_queue_empty():
-            write_pending_event(f"CONCURRENCY_LOW:actual={actual} expected={expected} queue=0")
-            log(f"CONCURRENCY_LOW:actual={actual} expected={expected} queue=0")
+            deficit = target - actual
+            write_pending_event(f"HARD_GATE:dispatch_required={deficit}:actual={actual} expected={expected} queue=0")
+            log(f"HARD_GATE:dispatch_required={deficit}:actual={actual} expected={expected} queue=0")
         else:
             actual = top_up_from_dispatch_queue(actual, target)
